Added a better conjure mode
Conjure mode now cycles through your modes based on a timer.

If you leave the lights off for 3 seconds and turn them back on, when
turning on they will activate the next mode in your bundle.

If you leave the lights off for 6 seconds, it will reset your lights
back to your first mode when you turn them back on.